win32: Set receive timeout os specific
For some reason the previous behaviour worked under windows xp
and windows 7, but failed under windows 10. As it turned out,
the function interface is different under windows and expects
a DWORD rather then a timeval struct.

So we use the DWORD argument under windows and the timeval for
posix now.

Anyhow, it is still unclear why only windows 10 shows the effect.

Signed-off-by: Soeren Grunewald <soeren.grunewald@desy.de>